requirements.txt
fastapi
uvicorn
httpx
python-dotenv
google-generativeai
asyncpg
sqlalchemy
twilio
httpx

